layout sign-in. Created sign-in-form component. Refactoring logo.
