+++ src/components/custom/logo/logo.module.scss
@@ -1,3 +1,7 @@
 .logo {
+  text-align: center;
+
+  img {
   cursor: pointer;
+  }
 }
