+++ src/components/custom/logo/logo.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 
 import styles from './logo.module.scss'
-import {ReactComponent as BrandLogo} from '../../../assets/brand-logo.svg'
+import BrandLogo from '../../../assets/logo.png'
 import {Link} from 'react-router-dom'
 
 const Logo: React.FC = () => {
   return (
     <div className={styles.logo} >
       <Link to="/">
-        <BrandLogo/>
+        <img src={BrandLogo} />
       </Link>
     </div>
   )
