+++ src/components/form/signInForm/signInForm.module.scss
@@ -1,5 +1,5 @@
-.signUp {
-  margin-top: 2rem;
+.signInForm {
+  padding-top: 2rem;
 
   input {
     padding: .2rem .4rem !important;
