+++ src/components/login/loginTabs/loginTabs.module.scss
@@ -1,7 +1,4 @@
 .loginTabs {
-  display: flex;
-  align-items: center;
-  justify-content: center;
   width: 100%;
 }
 
@@ -11,5 +8,5 @@
   justify-content: center;
   width: 100%;
   font-size: var(--login-page-tabs-font-size);
-  margin-right: 3rem;
+  margin-right: var(--login-page-tabs-gutter);
 }
