+++ src/components/login/signIn/SignIn.tsx
@@ -1,104 +1,22 @@
-import React, {useState} from 'react'
-import {
-  Form,
-  Input,
-  Button,
-  Checkbox
-} from 'antd';
-import {LockOutlined} from '@ant-design/icons';
+import React from 'react'
 
 import styles from './signIn.module.scss'
 import { LoginPageProps } from '../../../containers/LoginPageContainer';
+import SignInForm from '../../form/signInForm/SignInForm';
+import Logo from '../../custom/logo/logo';
 
 interface Props {
   signInWithEmail: LoginPageProps['signInWithEmail'],
 }
 
 const SingIn: React.FC<Props> = (props) => {
-  const [userData, setUserData] = useState({
-    email: '',
-    password: ''
-  })
-
-  const {email, password} = userData
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const {name, value} = event.target
-    setUserData({...userData, [name]: value})
-  }
-
-  const handleSubmit = async (event: React.FormEvent) => {
-    event.preventDefault()
-
-    props.signInWithEmail(email, password)
-  }
-
-  const clear = () => {
-    setUserData({
-      email: '',
-      password: ''
-    })
-  }
-
   return (
     <div className={styles.signIn}>
-      <Form
-        name='sign-in'
-        initialValues={{remember: true}}
-        onFinish={clear}
-        size='large'
-        preserve={false}
-      >
-        <Form.Item
-          name="email"
-          rules={[{
-            required: true,
-            message: 'Please input your email!',
-            type: 'email'
-          }]}
-        >
-          <Input
-            name="email"
-            placeholder="Email"
-            prefix={'@'}
-            allowClear
-            onChange={handleChange}
-            value={email}
-            autoFocus
-          />
-        </Form.Item>
+      <Logo />
 
-        <Form.Item
-          name="password"
-          rules={[{
-            required: true,
-            message: 'Please input your password!'
-          }]}
-        >
-          <Input.Password
-            name="password"
-            placeholder="Password"
-            prefix={<LockOutlined />}
-            onChange={handleChange}
-            value={password}
-          />
-        </Form.Item>
-
-        <Form.Item name="remember" valuePropName="checked">
-          <Checkbox>Remember me</Checkbox>
-        </Form.Item>
-
-        <Form.Item>
-          <Button
-            type="primary"
-            htmlType="submit"
-            onClick={handleSubmit}
-            className={styles.submitBtn}
-          >
-            Log in
-          </Button>
-        </Form.Item>
-      </Form>
+      <SignInForm
+        signInWithEmail={props.signInWithEmail}
+      />
     </div>
   )
 }
