+++ src/components/login/signIn/signIn.module.scss
@@ -1,7 +1,7 @@
 .signIn {
-  padding-top: 2rem;
-}
+  margin-top: var(--login-page-sign-in-mar-top);
 
-.submitBtn {
-  width: 100%;
+  & + div {
+    text-align: center;
+  }
 }
